Use housekeeper client from node-vtex-api
