+++ src/modules/sponsor/setEdition.ts
@@ -7,7 +7,7 @@ import log from '../../logger'
 import { default as switchAccount } from '../auth/switch'
 import { promptConfirm } from '../prompts'
 import { switchToPreviousAccount } from '../utils'
-import { getIOContext, options } from './utils'
+import { getIOContext, IOClientOptions } from '../utils'
 
 const promptChangeToSponsorAccount = async (sponsorAccount: string) => {
   const proceed = await promptConfirm(
@@ -22,7 +22,7 @@ const promptChangeToSponsorAccount = async (sponsorAccount: string) => {
 export default async (edition: string) => {
   const previousConf = conf.getAll()
   const previousAccount = previousConf.account
-  const sponsorClient = new Sponsor(getIOContext(), options)
+  const sponsorClient = new Sponsor(getIOContext(), IOClientOptions)
   const data = await sponsorClient.getSponsorAccount()
   const sponsorAccount = R.prop('sponsorAccount', data)
   if (!sponsorAccount) {
@@ -31,7 +31,7 @@ export default async (edition: string) => {
   if (previousAccount !== sponsorAccount) {
     await promptChangeToSponsorAccount(sponsorAccount)
   }
-  const sponsorClientForSponsorAccount = new Sponsor(getIOContext(), options)
+  const sponsorClientForSponsorAccount = new Sponsor(getIOContext(), IOClientOptions)
   await sponsorClientForSponsorAccount.setEdition(previousAccount, edition)
   log.info(`Successfully set new edition in account ${chalk.blue(previousAccount)}. You stil need to wait for the house keeper to update this account.`)
   await switchToPreviousAccount(previousConf)
